pyproject.toml
[tool.poetry]
name = "storybook"
version = "0.0.1"
description = "Kids storybook creator"
authors = ["Luca Simone <info@lucasimone.info>"]
license = "MIT"
readme = "README.md"

classifiers = [
  "Development Status :: 4 - Beta",
  "License :: OSI Approved :: MIT License",
  "Programming Language :: Python :: 3",
  "Programming Language :: Python :: 3.10",
]

[tool.poetry.dependencies]
python = "^3.10"
litellm = "^1.44.21"
diffusers = "^0.30.2"
pyyaml = "^6.0.2"
transformers = "^4.44.2"
numpy = "1.26.4"
torch = "2.2.2"
requests = "^2.32.3"
pillow = "^10.4.0"
instructor = "^1.4.1"
tqdm = "^4.66.5"

pydantic-settings = "^2.5.2"
[tool.poetry.group.dev.dependencies]
ruff = "^0.6.5"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
story = "storybook.main:main"
